Adds RRF_HOST_BUILD and fixes Hangprinter support
 - Only call serialUSB if SUPPORT_USB
 - Initializes some previously uninitialized variables.
 - Adds the missing Hangprinter registration so M669 K6 maps to a real kinematics.
   Instantiating the static descriptor and Create() helper now lets Move::SetKinematics construct HangprinterKinematics.
 - Fix warning in HangprinterKinematics.
   Replaces the variable-length stack buffers in CartesianToMotorSteps with HANGPRINTER_MAX_ANCHORS-sized arrays
   so GCC no longer flags zero-length writes when numAnchors is 0.
 - Implements new flex compensation for Hangprinter
 - Hangprinter: new quadratic forward transform
 - Hangprinter: all anchors drives control of all moves
   Hangprinter kinematics now advertises that all anchor drives control any Cartesian axis during normal moves, so async axis allocation includes motor C even when Z isn't mentioned.
   This mirrors the delta/CoreXY behavior.
 - Cache M569 values for expansion boards on main board.
   Hangprinter needs this for its can communication with ODrive on a separate can bus
 - Make M569.4 P<can_addrs> T<forces> allow a list of forces
   ... At least for Hangprinter
 - Hangprinter: Removes the M669 K-anchor to avoid name clash
 - Hangprinter: allow board address up to 49.0

+++ src/PrintMonitor/PrintMonitor.cpp
@@ -78,10 +78,10 @@ constexpr ObjectModelTableEntry PrintMonitor::objectModelTable[] =
 	// TODO Add enum about the last file print here (to replace lastFileAborted, lastFileCancelled, lastFileSimulated)
 	{ "layer",				OBJECT_MODEL_FUNC_IF(self->IsPrinting() && self->currentLayer != 0, (int32_t)self->currentLayer), 					ObjectModelEntryFlags::liveNotPanelDue },
 	{ "layerTime",			OBJECT_MODEL_FUNC_IF(self->IsPrinting() && self->currentLayer != 0, self->GetCurrentLayerTime(), 1), 				ObjectModelEntryFlags::liveNotPanelDue },
-	{ "pauseDuration",		OBJECT_MODEL_FUNC_IF(self->IsPrinting(), lrintf(self->GetPauseDuration())),											ObjectModelEntryFlags::liveNotPanelDue },
+	{ "pauseDuration",		OBJECT_MODEL_FUNC_IF(self->IsPrinting(), static_cast<int32_t>(lrintf(self->GetPauseDuration()))),					ObjectModelEntryFlags::liveNotPanelDue },
 	{ "rawExtrusion",		OBJECT_MODEL_FUNC_IF(self->IsPrinting(), ExpressionValue(self->gCodes.GetTotalRawExtrusion(), 1)),					ObjectModelEntryFlags::liveNotPanelDue },
 	{ "timesLeft",			OBJECT_MODEL_FUNC(self, 3),							 																ObjectModelEntryFlags::live },
-	{ "warmUpDuration",		OBJECT_MODEL_FUNC_IF(self->IsPrinting(), lrintf(self->GetWarmUpDuration())),										ObjectModelEntryFlags::live },
+	{ "warmUpDuration",		OBJECT_MODEL_FUNC_IF(self->IsPrinting(), static_cast<int32_t>(lrintf(self->GetWarmUpDuration()))),				ObjectModelEntryFlags::live },
 
 	// 1. 'file' members
 	{ "customInfo",			OBJECT_MODEL_FUNC(&self->customInfo, 0),						 													ObjectModelEntryFlags::none },
@@ -117,7 +117,7 @@ DEFINE_GET_OBJECT_MODEL_TABLE(PrintMonitor)
 
 int32_t PrintMonitor::GetPrintOrSimulatedDuration() const noexcept
 {
-	return lrintf((gCodes.IsSimulating()) ? gCodes.GetSimulationTime() + reprap.GetMove().GetSimulationTime() : GetPrintDuration());
+	return static_cast<int32_t>(lrintf((gCodes.IsSimulating()) ? gCodes.GetSimulationTime() + reprap.GetMove().GetSimulationTime() : GetPrintDuration()));
 }
 
 PrintMonitor::PrintMonitor(Platform& p, GCodes& gc) noexcept : platform(p), gCodes(gc), lastWarmUpDuration(0), isPrinting(false), heatingUp(false), paused(false), printingFileParsed(false)
@@ -462,7 +462,7 @@ float PrintMonitor::EstimateTimeLeft(PrintEstimationMethod method) const noexcep
 ExpressionValue PrintMonitor::EstimateTimeLeftAsExpression(PrintEstimationMethod method) const noexcept
 {
 	const float timeLeft = EstimateTimeLeft(method);
-	return (timeLeft > 0.0) ? ExpressionValue(lrintf(timeLeft)) : ExpressionValue(nullptr);
+	return (timeLeft > 0.0) ? ExpressionValue(static_cast<int32_t>(lrintf(timeLeft))) : ExpressionValue(nullptr);
 }
 
 // Return the estimated time until user interaction is required e.g. to change filament
@@ -474,7 +474,7 @@ ExpressionValue PrintMonitor::EstimateTimeToPause() const noexcept
 		const float timeToPause = slicerTimeToPause - (millis64() - whenSlicerTimeToPauseSet) * MillisToSeconds;
 		if (timeToPause > 0.0)
 		{
-			ret.SetInt(lrintf(timeToPause));
+			ret.SetInt(static_cast<int32_t>(lrintf(timeToPause)));
 		}
 	}
 	return ret;
